third_party/blink/web_tests/fast/forms/update-form-attribute-element-expected.txt

Test for the datalist element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test 1: Insert new input after form.
PASS form.elements.length is 1
PASS form.elements[0] is input
Test 2: Insert two forms with same id.
PASS form1.elements.length is 1
PASS form2.elements.length is 0
PASS form1.elements[0] is input
Test 3: Change form id.
PASS form.elements.length is 1
PASS form.elements[0] is input
Test 4: Order.
PASS form.elements.length is 5
PASS form.elements[0] is input1
PASS form.elements[1] is input2
PASS form.elements[2] is input3
PASS form.elements[3] is input4
PASS form.elements[4] is input5
PASS form.elements.length is 4
PASS form.elements[0] is input1
PASS form.elements[1] is input3
PASS form.elements[2] is input4
PASS form.elements[3] is input5
Test 5: Add new element with form attribute to the form.
PASS form.elements.length is 3
PASS form.elements[0] is input1
PASS form.elements[1] is input2
PASS form.elements[2] is input3
PASS successfullyParsed is true

TEST COMPLETE
